renamed registration dir to logistration
VAN-2
