+++ src/index.jsx
@@ -13,8 +13,8 @@ import configureStore from './data/configureStore';
 import AccountSettingsPage, { NotFoundPage } from './account-settings';
 import IdVerificationPage from './id-verification';
 import CoachingConsent from './account-settings/coaching/CoachingConsent';
-import LoginPage from './registration/LoginPage';
-import RegistrationPage from './registration/RegistrationPage';
+import LoginPage from './logistration/LoginPage';
+import RegistrationPage from './logistration/RegistrationPage';
 import appMessages from './i18n';
 
 import './index.scss';
