+++ src/index.scss
@@ -9,7 +9,7 @@ $fa-font-path: "~font-awesome/fonts";
 
 @import "./account-settings/style";
 @import "./id-verification/id-verification";
-@import "./registration/style";
+@import "./logistration/style";
 
 .word-break-all {
   word-break: break-all !important;
